Fix performance leak in sGetSimilarArticles
The function sGetSimilarArticles generates SQL code that does not perform well on shops with many articles (e.g. 300k), since the left joins unneccessary create multiple lines. In an example case the result with 350k articles, 500  categories (approx 3 categories for 1 article - its a book store - ) this query generates over 2,1 million lines  and runs on a quite big dedecated server for 7-15sek, since the big result tables needs to be written to disc my the database.

To group and sort 2,1 million lines for 4 random lines picket by a simple relevance is not good.

When a crawler comes by a shop and crawls many unavailable products (ignoring the sitemap) the server is flooded by long running queries and begins to lag.

The idea here is, to split those 3 relevances into independent queries und unionize the result. This yields the exact same four randomized results (in terms of relevance, whilst of course not the exact same 4, since the DB does not have to order the result).